+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 CHANGELOG
 ---------
-- **Unreleased**:
+- **Unreleased**: 4.4.0
+  - [PR #469](https://github.com/caxlsx/caxlsx/pull/469) Add default theme file to Excel package.
   - [PR #475](https://github.com/caxlsx/caxlsx/pull/475) Use timecop to fix transient time failure in tests
   - [PR #474](https://github.com/caxlsx/caxlsx/pull/474) Add Windows and MacOS to the CI.
   - [PR #474](https://github.com/caxlsx/caxlsx/pull/474) Fix local image file MIME type detection on Windows.